LLVM tools: Added bisection of LLVM
Bisects LLVM based using a JSON file and the last good version and the
last bad version. The number of tryjobs that are launched in the between
the two (good, bad) depends on the 'parallel' command line argument.

Added support for modifying a tryjob's status.

Added support for removing, relaunching, or adding a tryjob.

BUG=None
TEST=Ran the script with 'start_rev' as 367590 and 'end_rev' as 367622
and a JSON file that did not exist. Successfully created tryjobs. Ran,
the script again after a few hours and also successfully created another
set of tryjobs that were not in the 'jobs' list.

Change-Id: I71a6d74f62c6f6fc212209c1faa73e729912b086
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1761469
Reviewed-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Salud Lemus <saludlemus@google.com>