crosperf: Add support for using aquarium in Telemetry.
We want to test the story aquarium in rendering.desktop and see
the results comparison with graphics_WebGLAquarium in autotest.
This patch adds support to filter out the most important metric to
display in the report for the two aquarium stories (with different
number of fish). Also fix the logic of reading histogram results
when there's None value in the list.

Also add the test to toolchain nightly tests.

BUG=None
TEST=crosperf --board=nami with the two benchmarks

Change-Id: Ia1aa6bf796bee78bf6199332a07f4b96aa084845
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1779248
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Tiancong Wang <tcwang@google.com>
Tested-by: Tiancong Wang <tcwang@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>