LLVM tools: Added a simple custom script example
This custom script determines what to set the tryjob's 'status' to based
off of the script's exit code.

BUG=None
TEST=Ran the 'update_tryjob_status.py' script with `--custom_script`
and `--set_status custom_script` option. Used the absolute path to this
script for `--custom_script`. The tryjob's 'status' was updated
correctly based off of the exit code by the custom script.

Change-Id: Ib4fe58406ba6f9c9cb0cdc9dcc3e7d5e9bb18fb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1783838
Reviewed-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Salud Lemus <saludlemus@google.com>