llvm_tools: support grabbing google3's unstable LLVM
Google3 integrates LLVM pretty often -- using that as a foundation for
testing might be a good idea. This CL enables us to pull the revision of
their unstable LLVM.

As an added bonus, this will also have us print a message to stderr when
we're syncing LLVM for the first time. Otherwise, it may seem to a user
that our script got hung up on something while that first clone happens.
:)

BUG=chromium:914081
TEST=Unittests. Ran the new get_llvm_hash functionality.

Change-Id: I3a7088c25dbd7169ca31373daf0c46fb28f46a68
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1793906
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>