toolchain-utils: fix an AFE locking bug and reformat status
There is a bug in afe_lock_machine that if a machine is not in toolchain
pool but still managed by afe, locking/unlocking will not perform
correctly. This CL fixes this issue.

Also this CL reformated the output of the machine status and we are now
calling the print status function by default when getting machines
status. An example looks like this:

Machine (Board)					Status
---------------					------
chromeos4-row6-rack2-host13.cros (lulu)		unlocked

TEST=Tested with machines in different pools managed by afe.
BUG=chromium:1006434

Change-Id: I3213ae3ca6c153361753ca91686891cc44d8a5a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1846261
Tested-by: Zhizhou Yang <zhizhouy@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Zhizhou Yang <zhizhouy@google.com>