crosperf: fix a download image bug
This is a fix to a bug introduced by
https://chromium-review.googlesource.com/1865958. GetXbuddyPath()
function is also called in experiment_file and it's for re-writing the
fields to output. So at that time, we don't need to really check whether
downloading debug again.

BUG=chromium:1010294
TEST=tested with kevin locally.

Change-Id: Ic7daf6d1242f012bdef803faca355df0365f2cdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1895979
Tested-by: Zhizhou Yang <zhizhouy@google.com>
Auto-Submit: Zhizhou Yang <zhizhouy@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Denis Nikitin <denik@chromium.org>