crosperf: Make StopUI/StartUI more robust
There was a bug when crosperf failed due to StopUI failure.
This happened when in a preceding run crosperf exits with an exception
in DeviceSetup (for example Keyboard interrupt) and fails to call
StartUI.

Current change fixes this problem and includes a unittest case testing
the exception case.

BUG=None
TEST=unittest and HW test on DUT with stopped ui pass

Change-Id: Id6c69aebefe21b12ec7ee3a7c7f9dff92d143908
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1941036
Commit-Queue: Denis Nikitin <denik@chromium.org>
Tested-by: Denis Nikitin <denik@chromium.org>
Reviewed-by: Zhizhou Yang <zhizhouy@google.com>