Crosperf: Do not image DUT in Skylab mode
We should not image DUT with certain image when the test is going to be
launched via 'skylab create-test'. This is because the DUT may be
occupied by other tasks before Crosperf acutally get it assigned, and
the first step after the test starts is to flash image.

This patch is a temproray solution to ensure no bad behavior happens on
the DUT before testing.

TEST=Tested wth skylab run, w/o caching.
BUG=chromium:984790

Change-Id: Ie2f439d5777813905bdca5f4a2d1e401df2dc83e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1952172
Tested-by: Zhizhou Yang <zhizhouy@google.com>
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Commit-Queue: Zhizhou Yang <zhizhouy@google.com>
Auto-Submit: Zhizhou Yang <zhizhouy@google.com>