crosperf: fix skylab telemetry_Crosperf tests launching command
Current command generated from suite_runner for telemetry_Crosperf run
is not correct, and this patch will fix it. Also added an unittest to
verify it in the future.

It is unfortunate that we currently still cannot run test with args
successfully in skylab, but according to the args passed to skylab, this
change should work.

TEST=Checked command line to launch skylab create-test.
BUG=None

Change-Id: I05f0f4c60e04d19e2a2637192023829d4dd9f48b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1963355
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Zhizhou Yang <zhizhouy@google.com>
Tested-by: Zhizhou Yang <zhizhouy@google.com>
Auto-Submit: Zhizhou Yang <zhizhouy@google.com>