afdo_metadata: Add entries for kernel 4.14 and 4.19
Adds two entries for kernel 4.14 and 4.19 in the metadata file,
which will be needed by the kernel PUpr to update ebuild files.
Using the same name as the ebuild files in this initial upload,
and will be updated by kernel AFDO verification builders later.

BUG=chromium:1026315
TEST=None

Change-Id: Ic2ff8f1310d691d72208f70bee619eee4b4093c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2067155
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: David Burger <dburger@chromium.org>
Commit-Queue: Tiancong Wang <tcwang@google.com>