compiler_wrapper: Add "-fcommon" to clang invocations
Clang ToT has switched to "-fno-common" as default to match GCC 10 which
breaks many packages. Upgrading to ToT versions of many packages does not
fix the problem as packages upstream have not fixed them yet.
Pass "-fcommon" in compiler wrapper to workaround this for now.
This does not change the current clang behavior which defaults to
"-fcommon" so this is a NFC.

BUG=chromium:1060413
TEST=packages build with ToT clang; NFC for current clang

Change-Id: I5a40e502db201615f45a7e4720105bebef2f2b0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2102905
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>