compiler_wrapper: Stop setting CCACHE_BASEDIR
CCACHE_BASEDIR used to be set to sysroot so that ccache can reuse
compilation results across different boards. This is not useful
on builders and also less useful for board builds since we set
closer to target CFLAGS now.
It also breaks use of -fmacro-prefix-map option in clang/gcc.
So rip it out and simplify the rest of code.

BUG=chromium:983759
TEST=go test

Change-Id: Ia99ca994c976ee1137c3d904a78f7e25931c70f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2103170
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>