cros_utils: Do not sort keys by default
Add a sort option NO_SORT in TableGenerator and make it default.

Before we had rows in the table sorted by key-name which mixed up the
test scores with cpufreq_*, cputemp_* data and made it hard to locate
the most important information.
The change solves the problem by leaving the keys in the original order
which will guarantee that retval and cpu stats will always be at the
end of the table.

BUG=None
TEST=rendering.desktop and loading.desktop show scores in the beginning
of the table.

Change-Id: I5b2d75db20549eae264db9ca809ccd634e743510
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2108572
Tested-by: Denis Nikitin <denik@chromium.org>
Commit-Queue: Denis Nikitin <denik@chromium.org>
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>