compiler_wrapper: don't write -Werror files on still-failed builds
Not `return`ing here was troublesome for our monitoring infra, since it
meant that _any_ clang error that mentioned '-Werror', with this support
turned on, would turn into a json file. With configure steps and the
like, this probably isn't the behavior that we want.

BUG=None
TEST=unittests

Change-Id: I0669c5bd9473b687c34c23f284433a21e8a1d88c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2132233
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>