cros_utils: import tiny_render + some (new) tests from google3
Purely plaintext emails are nice, but HTML can be way easier to deal
with. `tiny_render` is a super simple way of being able to render text
and HTML from the same structure, so we can send out descriptive emails
accessible to people with various kinds of email clients.

tiny_render was imported with slight modifications from
//wireless/android/llvm/monitoring/commits/tiny_render.py . Tests are
new.

BUG=chromium:1046988
TEST=unittests

Change-Id: Ic94064d0125d3e7655498c9c2b8501f1448420d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2169009
Reviewed-by: Tiancong Wang <tcwang@google.com>
Tested-by: George Burgess <gbiv@chromium.org>