lock_machine: shout when our `swarming` command dies
When I tried to run crosperf without proper `swarming.py`
authentication, crosperf said the machine I wanted to use wasn't
available in skylab. It'd be better if we reported that we couldn't tell
if the machine was/wasn't in skylab due to the tool we're using failing.

Looks like this script can fail for a number of other reasons, too, so
throwing our hands up is likely going to often be our best bet.

BUG=None
TEST=ran without auth; exception was thrown. ran with; things worked.

Change-Id: I4d0e4ec70eddb3a7836686eef819e6175a730756
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2177063
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>