Replace 'veyron-minnie' with 'veyron_minnie' for vanilla image lookup.
The recipe builders use hyphens in the board archives, so when we get
the trybot image from the recipe builder archives, it has a hyphen
rather than an underscore.  We then use the trybot image name to find
the corresponding vanilla image.  But the vanilla images are in archives
that are still using the underscore in the board name (for veyron_minnie).
So we need to update the trybot name to use the underscore again before
looking for the vanilla image.

BUG=chromium:1077618
TEST=None

Change-Id: I01057a8546e87af0157514003cd63729853a1552
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2177651
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>