crosperf: Include PID in topstats
Top statistics was showing commands which could combined multiple
processes. To include PID in topstats we need to split the commands into
separate processes.
The process PID is appended to the command name in topstats. The top
chrome process is the renderer process running the benchmark.

The list size depends on the number of non-chrome processes which we
limit to 5. For example with 10 chrome processes in the top with 10 following
non-chrome processes the list will show 15 entries.

BUG=None
TEST=Tested on eve, bob and cheza.

Change-Id: Ibf1e61c8cb522aba13cd51a590bb7e24597f66a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2227626
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Denis Nikitin <denik@chromium.org>