Symlink to android_llvm_next_flags.go in toolchain/llvm_android
This is a downstream/Android-only change to symlink to
android_llvm_next_flags.go in toolchain/llvm_android.  It, as the name
suggests, has custom flags for Android's LLVM-next compiler.  This file
is maintained in the llvm_android project to track configurations for
the Android toolchain in one project.

Test: compiler_wrapper/build.py for Android and validate that flags are
passed correctly.

Change-Id: I7a535aa6b39119864024495c33427504cbcb3c64