llvm_tools: fix up naming
This replaces `master` with `main` where possible, per the linked bug.
Since upstream LLVM has a `master` branch, and there appears to be
general support in FOSS for swapping to a different naming convention
(`main`), this also makes LLVM's upstream branch name into a constant
that we can easily flip if needed.

BUG=chromium:1099035
TEST=Unittests

Change-Id: I0844cfb4fec00f761e39b0443299daa918fa37ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2265047
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>