compiler_wrapper: remove -grecord-gcc-switches
-grecord-gcc-switches causes the compiler and compiler flags to be
stored in the produced debug information. This causes Goma builds to
be non-deterministic, because the compiler path differs depending
on whether the compiler was run locally or on Goma. This, in turn,
causes cache pollution and needless work. Since we are not in
fact using the information stored by -grecord-gcc-switches, this
change disables it. We can re-enable it when we want to start
using the information, provided we come up with another way to
avoid the issues described earlier.

BUG=chromium:1103065
TEST=Build an object file in chromeos-chrome and check that info is absent

Change-Id: Ib8022a3a1baf2e4b5d3d9d1453deb67f539baff4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2290624
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Tiancong Wang <tcwang@google.com>