rust_uprev: emerge rust instead of ebuilding it
if `ebuild merge` fails, the build directory might not be in a great
state where we can skip recompilation when we try again. emerge'ing is
probably what we actually want to do here, since that's representative
of what builders and users are going to do when they get the new rust
uprev.

BUG=chromium:1136579, chromium:1112551
TEST=Ran the script

Change-Id: I84f12769e32fada3f858a2889f0ead7555b58256
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2462928
Reviewed-by: Tiancong Wang <tcwang@google.com>
Tested-by: George Burgess <gbiv@chromium.org>