compiler_wrapper: report parent processes' info in warnings tarballs
The bug I'm chasing down right now depends on environment variables, and
could be due to a `make` invocation in one of N places. Having this
context available to me is useful.

It seriously pollutes the JSON we dump, but use of `jq` that I've put in
the doc about this makes that a non-issue.

BUG=None
TEST=Ran on stress-ng. parent process info was dumped.

Change-Id: Ib4d3b6c9887b4cf2a38472ef299c81af27e9a617
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2481829
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>