llvm_tools: detect cros checkout for default chroot_path
This CL improves the assumptions made when selecting the default chroot
path in update_chromeos_llvm_hash.py. Previously, it assumed the chrome
source tree was installed to ~/chromiumos which didn't work for people
with multiple chroots or with different naming. Now we detect whether
or not we are inside of the cros source tree and use that cros checkout
if so (otherwise defaulting back to old logic).

BUG=None
TEST=./update_chromeos_llvm_hash.py produced CL:2912507 without
specifiying chroot_path manually

Change-Id: Idb6dfc18a6f7a95c36eba4209dc43e08e373d6e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/2912956
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>