llvm_tools: use upstream revert_checker
Sourcing this properly from LLVM is awkward, since at the least, we'd
need to keep an LLVM git repo around. On top of that, we'd need to come
up with some way to pin this at a specific version (since the evolution
of LLVM ideally shouldn't impact the functionality of bits here).

Rather than dealing with all of that, let's just copy this from the
repo and declare victory. Changes should be super rare anyway. :)

BUG=chromium:194731505
TEST=./nightly_revert_checker_test.py

Change-Id: I6179cae57d01bee7cae7883c93f71cb9c2f1f3d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3068861
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>