Add upstream-workon.bash script
upstream-workon is a script which aids toolchain developers by
keeping the local, fully patched source in some non-ephemeral
local directory inside the chroot.

At the moment, it's still in a prototype phase (the end goal
would be to make it quite a bit safer, and also move it out of
bash).

This is the first commit for the script, meant to act as a
base for future work.

BUG=b:186803200
TEST=mkdir ~/llvm; upstream_workon.bash init sys-devel/llvm ~/llvm
TEST=upstream_workon.bash build sys-devel/llvm
TEST=upstream_workon.bash install sys-devel/llvm
TEST=upstream_workon.bash link sys-devel/llvm ~/llvm_src_tree

Change-Id: Ifa1b1520d621cbe8b435d4ebb1ae71dc28548fcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3068075
Commit-Queue: Jordan R Abrahams <ajordanr@google.com>
Tested-by: Jordan R Abrahams <ajordanr@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>