Excise cros_logging module
The cros_logging module is deprecated and all usages are being removed.
The lint referred to by the comment removed in this CL is no longer a
thing, so remove this mention of the deprecated module from the tree.

BUG=b:191490453
TEST=None

Change-Id: Ifcd4cb18e87fdbd3a92747b0fdbac434d51fd055
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3161649
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Auto-Submit: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>