compiler_wrapper: fix use-llvm-next test
The history of this test looks weird. We've had it intermittently
enabled/disabled depending on whether we're passing flags specifically
to llvm-next versions of clang. At the moment, it's failing due to a
lack of `-Wl,-z,nostart-stop-gc`, which was removed in a previous CL.

BUG=None
TEST=go test

Change-Id: I4b1ab9e5d39ba0e5575f89f51e4e83afba24a025
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3162667
Reviewed-by: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>