rust_uprev_test: mock equery w commands
rust_uprev uses `equery w` to find the ebuild file for a package.
This depends on filesystem state and so should not affect the test.
This change mocks these calls so that they return specified values
instead of actually running commands and accessing the filesystem.

BUG=None
TEST=Ran the tests outside the chroot, where equery does not exist

Change-Id: Id48a0e063f9bf112888d378d549420f008fdcb4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3182629
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Michael Benfield <mbenfield@google.com>
Commit-Queue: Bob Haarman <inglorion@chromium.org>