patch_sync: Base code, version control managing
This is the base code addition for the patch_sync project. This
project attempts to synchronize LLVM patches and ChromeOS. This
code is not meant to run, but simply be the starting point so
that we break up the large code dump.

This code is meant to compile, but is not meant to be run. It
only adds the version control code which manages `repo`
interactions.

Additionally, this adds the full Cargo.toml (at least for
the current patch stack).

BUG=b:209493133
TEST=cargo build

Change-Id: I903edba6e633ec53a7fa4d2e7b0c2aac1094933a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3339404
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Jordan Abrahams <ajordanr@google.com>
Commit-Queue: Jordan Abrahams <ajordanr@google.com>