llvm_tools: Update get_upstream_patch schema
The get_upstream_patch.py program is using the old schema for
PATCHES.json, and appending with that old schema to the new
format. This commit fixes this issue.

Additionally, this ensures that the keys are sorted on write
to prevent line-diffs from getting out of sync.

BUG=b:209493133
TEST=get_upstream_patch $args --platform chromiumos --platform android

Change-Id: I0f96843b0134daa27574a94f583efdd27f394a3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3352574
Reviewed-by: Christopher Di Bella <cjdb@google.com>
Commit-Queue: Jordan Abrahams <ajordanr@google.com>
Tested-by: Jordan Abrahams <ajordanr@google.com>