patch_sync: Add better command debugging
This commit allows the git commands to print out
what the command was that they failed at. This
is pretty helpful in debugging what went wrong.

We don't capture the repo output ever, so that
can remained piped to the terminal. But we
need to do some trickery for the git cmd.

BUG=b:209493133
TEST=None

Change-Id: I389257fef1e3bf394fb4013588df6c78e83b733a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3379478
Reviewed-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Michael Benfield <mbenfield@google.com>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>