patch_sync: Bugfix for ebuild finding
Previously, the find_ebuild function could return the original
ebuild instead of the llvm symlink. This would lead to incorrect
uprevs, depending on how the file system returned the order of
the files.

Adds a test case to prevent this from happening in the future.

BUG=b:209493133
TEST=cargo test

Change-Id: Ia47ecee2a9c5b6ce0559e316c5227f70dcf87833
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3379482
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>