patch_sync: Toggle CQ for CrOS, Android
Currently, patch sync will always run CQ for CrOS, but never for
Android. This was because the chrotomation bot didn't have the
proper permissions to run the needed Presubmit-Ready checks.

Now the bot does in fact have permission to enable the Presubmit
check on Android, but some users may not have this permission when
running locally. This commit introduces the ability to toggle
running the presubmit/CQ checks early through the --disable-cq flags
(by default it's assumed that running these checks is desirable).

BUG=None
TEST=patch_sync transpose --disable-cq <...>
TEST=patch_sync transpose <...>

Change-Id: I6185b98aa4394ba22f8433541885a450855cfbb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3451037
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>