lock_machine: specify python3 for swarming
swarming.py was being called with python2 which no longer exists in
chrotomation. This was later changed to make the invokaction call
swarming as an exexutable, however, this lead swarming.py to then
invoke the default `python` executable which also doesn't exist
in chrotomation and so a similar problem ensued.

BUG=b:221777277
TEST=manually verified

Change-Id: Ie6537b00a25fbc12cdeb95ada92d8095a609b291
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3553430
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>